requirements.txt
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
idna==2.10
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
pony==0.7.14
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
vk-api==11.9.1
